Separate MapBox utility functions into utils file
